+++ src/container/RestaurantsContainer.jsx
@@ -1,5 +1,6 @@
 import { useSelector } from 'react-redux';
-import Restaurants from './Restaurants';
+
+import Restaurants from '../presentational/Restaurants';
 
 export default function RestaurantsContainer() {
   const { restaurants } = useSelector((state) => ({
